CLI: refactor, move CLI error its own module
