+++ lib/uniffi-zcash-cli/src/cli.rs
@@ -1,9 +1,11 @@
-use std::{error::Error, fmt::Display};
-
-use clap::{builder::ValueParser, parser::MatchesError, Arg, ArgMatches, Command};
+use clap::{builder::ValueParser, Arg, ArgMatches, Command};
 use strum::VariantNames;
 
 use crate::SupportedLangs;
+
+use self::error::CLIError;
+
+mod error;
 
 pub fn get_matches() -> ArgMatches {
     Command::new("UniFFI Zcash CLI")
@@ -56,65 +58,6 @@ pub fn validator_regex(regex: &'static str, err_msg: &'static str) -> ValueParse
             false => Err(format!("Value \"{}\" is not matching format: {}", input, err_msg).into()),
         }
     })
-}
-
-#[derive(Debug)]
-pub struct CLIError {
-    message: String,
-}
-
-impl Error for CLIError {}
-
-impl Display for CLIError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&self.message)
-    }
-}
-
-impl From<&str> for CLIError {
-    fn from(value: &str) -> Self {
-        CLIError {
-            message: value.to_string(),
-        }
-    }
-}
-
-impl From<String> for CLIError {
-    fn from(value: String) -> Self {
-        Self { message: value }
-    }
-}
-
-impl From<std::io::Error> for CLIError {
-    fn from(value: std::io::Error) -> Self {
-        CLIError {
-            message: value.to_string(),
-        }
-    }
-}
-
-impl From<MatchesError> for CLIError {
-    fn from(value: MatchesError) -> Self {
-        Self {
-            message: value.to_string(),
-        }
-    }
-}
-
-impl From<fs_extra::error::Error> for CLIError {
-    fn from(value: fs_extra::error::Error) -> Self {
-        Self {
-            message: value.to_string(),
-        }
-    }
-}
-
-impl From<handlebars::RenderError> for CLIError {
-    fn from(value: handlebars::RenderError) -> Self {
-        Self {
-            message: value.to_string(),
-        }
-    }
 }
 
 pub type CLIResult<T> = Result<T, CLIError>;
